Move all fill-related C++ code to separate dir
To avoid swamping the lib directory with all the new headers
and source files that provide fill-tool-specific functionality,
these are now moved to a subdirectory.
