+++ lib/fill/gap_closing_fill.cpp
