+++ lib/mypaintlib.hpp
@@ -23,10 +23,10 @@
 #include "colorchanger_crossed_bowl.hpp"
 #include "gdkpixbuf2numpy.hpp"
 #include "fastpng.hpp"
-#include "fill_constants.hpp"
-#include "fill_common.hpp"
-#include "floodfill.hpp"
-#include "gap_detection.hpp"
-#include "blur.hpp"
-#include "morphology.hpp"
+#include "fill/fill_constants.hpp"
+#include "fill/fill_common.hpp"
+#include "fill/floodfill.hpp"
+#include "fill/gap_detection.hpp"
+#include "fill/blur.hpp"
+#include "fill/morphology.hpp"
 #include "brushsettings.hpp"
