+++ setup.py
@@ -578,17 +578,17 @@ def get_ext_modules():
         'lib._mypaintlib',
         [
             'lib/mypaintlib.i',
-            'lib/fill_common.cpp',
-            'lib/fill_constants.cpp',
-            'lib/floodfill.cpp',
-            'lib/gap_closing_fill.cpp',
-            'lib/gap_detection.cpp',
-            'lib/blur.cpp',
-            'lib/morphology.cpp',
             'lib/gdkpixbuf2numpy.cpp',
             'lib/pixops.cpp',
             'lib/fastpng.cpp',
             'lib/brushsettings.cpp',
+            'lib/fill/fill_common.cpp',
+            'lib/fill/fill_constants.cpp',
+            'lib/fill/floodfill.cpp',
+            'lib/fill/gap_closing_fill.cpp',
+            'lib/fill/gap_detection.cpp',
+            'lib/fill/blur.cpp',
+            'lib/fill/morphology.cpp',
         ],
         swig_opts=mypaintlib_swig_opts,
         language='c++',
